refact: daily learning
