+++ src/Domains/Connectors/Hermes/Actions/FetchDailyLearningContextAction.php
@@ -4,54 +4,22 @@ declare(strict_types=1);
 
 namespace Kanvas\Connectors\Hermes\Actions;
 
-use Illuminate\Support\Facades\Log;
 use Kanvas\Connectors\Hermes\SshClient;
-use Kanvas\Intelligence\Agents\Models\AgentDeployment;
-use RuntimeException;
-use Throwable;
+use Kanvas\Intelligence\AgentRuntime\Actions\BaseFetchDailyLearningContextAction;
+use Kanvas\Intelligence\AgentRuntime\SshClient as BaseSshClient;
+use Kanvas\Intelligence\Agents\Models\AgentMachine;
+use Override;
 
-/**
- * Read the agent's current MEMORY.md so the summarize prompt can inject it
- * as "facts already known — don't re-emit." Symmetric counterpart to
- * PushDailyLearningContextAction; same path resolution, opposite direction.
- *
- * Returns '' (not throws) when the file is missing or unreadable — empty
- * string is the documented "nothing to share" answer per the runtime
- * contract, so the caller doesn't need a try/catch.
- */
-class FetchDailyLearningContextAction
+class FetchDailyLearningContextAction extends BaseFetchDailyLearningContextAction
 {
-    public function __construct(
-        protected readonly AgentDeployment $deployment,
-    ) {
-    }
-
-    public function execute(): string
+    #[Override]
+    protected function createSshClient(AgentMachine $machine): BaseSshClient
     {
-        $machine = $this->deployment->machine;
-        if ($machine === null) {
-            Log::warning('Hermes daily-learning fetch skipped: deployment has no machine', [
-                'deployment_id' => $this->deployment->getId(),
-            ]);
-            return '';
-        }
-
-        $ssh = SshClient::fromMachine($machine);
-
-        try {
-            return $this->safeReadFile($ssh, $this->resolveMemoryPath());
-        } catch (Throwable $e) {
-            Log::warning('Hermes daily-learning fetch failed', [
-                'deployment_id' => $this->deployment->getId(),
-                'error' => $e->getMessage(),
-            ]);
-            return '';
-        } finally {
-            $ssh->disconnect();
-        }
+        return SshClient::fromMachine($machine);
     }
 
-    private function resolveMemoryPath(): string
+    #[Override]
+    protected function resolveMemoryPath(): string
     {
         $home = $this->deployment->home_directory !== ''
             ? $this->deployment->home_directory
@@ -60,16 +28,9 @@ class FetchDailyLearningContextAction
         return rtrim($home, '/') . '/.hermes/memories/MEMORY.md';
     }
 
-    /**
-     * Missing file is the first-push case — return empty rather than throw.
-     * Other SSH errors propagate to the catch in execute().
-     */
-    private function safeReadFile(SshClient $ssh, string $path): string
+    #[Override]
+    protected function runtimeName(): string
     {
-        try {
-            return $ssh->readFileAsUser($path);
-        } catch (RuntimeException) {
-            return '';
-        }
+        return 'Hermes';
     }
 }
